pyproject.toml
[tool.poetry]
name = "omniknight"
version = "0.1.0"
description = "By the Omniscience."
authors = ["Nasy <nasyxx+bot@gmail.com>"]
license = "GPL-3.0+"

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = {version = "^11.1",allows-prereleases = true}

[tool.poetry.dev-dependencies]

[tool.black]
line-length = 79

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
